pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "semfire"
version = "0.2.0"
authors = [{ name = "Edward Joseph", email = "author@example.com" }]
description = "SemFire (Semantic Firewall): detect advanced AI deception, including in-context scheming and multi-turn manipulative attacks."
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]
keywords = ["semfire", "llm", "security", "prompt injection", "semantic firewall", "ai safety", "deception detection"]
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "accelerate>=0.20.0",
    "openai>=0.27.0",
    "colorama>=0.4.6",
    "python-dotenv>=0.21.0"
]

[project.urls]
"Homepage" = "https://github.com/josephedward/R.A.D.A.R."
"Repository" = "https://github.com/josephedward/R.A.D.A.R."
"Bug Tracker" = "https://github.com/josephedward/R.A.D.A.R./issues"

[project.optional-dependencies]
api = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.22.0"
]
demo = [
    "streamlit>=1.20.0"
]
dev = [
    "pytest",
    "pytest-cov",
    "PyGithub>=1.55",
    "python-dotenv>=0.21.0",
    "requests>=2.28.0",
    "selenium>=4.8.0",
    "webdriver-manager>=3.8.5",
]

[project.scripts]
semfire = "cli:main"
aegis = "cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
py-modules = ["cli", "semantic_firewall", "semfire", "config_menu"]

[tool.setuptools.packages.find]
where = ["src"]
namespaces = true
